Extracted ImageList array to own file, added orientation for full size images, amde background more opaque for full screen images, extracted scroll to top button to own component
