+++ src/Archive/Archive.scss
@@ -22,17 +22,6 @@
     text-align: center;
 }
 
-.collage-scroll-icon.MuiIconButton-root {
-    background-color: $school-green;
-    color: $main-white;
-    display: block;
-    margin: auto;
-    &:hover {
-        background-color: $school-green;
-        box-shadow: 0px 0px 10px $main-white;
-    }
-}
-
 .archivePageContent {
     margin: 5rem 2rem;
     display: flex;
