+++ src/Archive/ImageGrid.tsx
@@ -5,20 +5,9 @@ import { GridList, GridListTile, GridListTileBar } from "@material-ui/core";
 import Image from "material-ui-image";
 import './Archive.scss';
 import { COLORS } from '../Utils/COLORS';
+import { tileData, ArchiveCategory } from './ImageList';
 
-export enum ArchiveCategory {
-  all = 0,
-  outreach,
-  sprintReview,
-  robot,
-}
 
-interface Tile {
-  img: string;
-  title: string;
-  category: ArchiveCategory;
-  featured: boolean;
-}
 
 const useStyles = makeStyles((theme: Theme) =>
   createStyles({
@@ -35,6 +24,7 @@ const useStyles = makeStyles((theme: Theme) =>
       height: "100%",
       transform: 'translateZ(0)',
       scrollBehavior: "smooth",
+      cursor: "pointer",
     },
     titleBar: {
       background:
@@ -43,138 +33,9 @@ const useStyles = makeStyles((theme: Theme) =>
   }),
 );
 
-const tileData: Tile[] = [
-  {
-    img: '/ReturningMembers.jpg',
-    title: 'Returning Members',
-    category: ArchiveCategory.all,
-    featured: true,
-  },
-  {
-    img: '/Robot1.jpg',
-    title: 'Robot',
-    category: ArchiveCategory.robot,
-    featured: false,
-  },
-  {
-    img: '/Robot2.jpg',
-    title: 'Robot',
-    category: ArchiveCategory.robot,
-    featured: false,
-  },
-  {
-    img: '/Outreach/ArduinoWksp-6.jpg',
-    title: 'Arduino WorkSpace',
-    category: ArchiveCategory.outreach,
-    featured: true,
-  },
-  {
-    img: '/Outreach/ArduinoWksp-5.jpg',
-    title: 'Arduino WorkSpace',
-    category: ArchiveCategory.outreach,
-    featured: false,
-  },
-  {
-    img: '/Outreach/ArduinoWksp-2.jpg',
-    title: 'Arduino WorkSpace',
-    category: ArchiveCategory.outreach,
-    featured: false,
-  },
-  {
-    img: '/Outreach/ArduinoWksp-8.jpg',
-    title: 'Arduino WorkSpace',
-    category: ArchiveCategory.outreach,
-    featured: true,
-  },
-  {
-    img: '/Outreach/BobsBash1-2.jpg',
-    title: 'Bobs Bash',
-    category: ArchiveCategory.outreach,
-    featured: true,
-  },
-  {
-    img: '/Outreach/BobsBash1-1.jpg',
-    title: 'Bobs Bash',
-    category: ArchiveCategory.outreach,
-    featured: false,
-  },
-  {
-    img: '/Outreach/DiscoveryPlace2.jpg',
-    title: 'Discovery Place',
-    category: ArchiveCategory.outreach,
-    featured: false,
-  },
-  {
-    img: '/Outreach/DiscoveryPlace1.jpg',
-    title: 'Discovery Place',
-    category: ArchiveCategory.outreach,
-    featured: true,
-  },
-  {
-    img: '/Outreach/Elementary2.jpg',
-    title: 'Elementary',
-    category: ArchiveCategory.outreach,
-    featured: true,
-  },
-  {
-    img: '/Outreach/Elementary1.jpg',
-    title: 'Elementary',
-    category: ArchiveCategory.outreach,
-    featured: false,
-  },
-  {
-    img: '/Outreach/NCScience2.jpg',
-    title: 'NCScience',
-    category: ArchiveCategory.outreach,
-    featured: false,
-  },
-  {
-    img: '/Outreach/NCScience1.jpg',
-    title: 'NCScience',
-    category: ArchiveCategory.outreach,
-    featured: true,
-  },
-  {
-    img: '/SprintReviews/EndofSprint2-1.jpg',
-    title: 'Sprint Reviews',
-    category: ArchiveCategory.sprintReview,
-    featured: true,
-  },
-  {
-    img: '/SprintReviews/EndofSprint2-2.jpg',
-    title: 'Sprint Reviews',
-    category: ArchiveCategory.sprintReview,
-    featured: false,
-  },
-  {
-    img: '/SprintReviews/EndofSprint2-3.jpg',
-    title: 'Sprint Reviews',
-    category: ArchiveCategory.sprintReview,
-    featured: false,
-  },
-  {
-    img: '/SprintReviews/EndofSprint2-4.jpg',
-    title: 'Sprint Reviews',
-    category: ArchiveCategory.sprintReview,
-    featured: true,
-  },
-  {
-    img: '/SprintReviews/EndofSprint2-5.jpg',
-    title: 'Sprint Reviews',
-    category: ArchiveCategory.sprintReview,
-    featured: false,
-  },
-  {
-    img: '/SprintReviews/EndofSprint2-7.jpg',
-    title: 'Sprint Reviews',
-    category: ArchiveCategory.sprintReview,
-    featured: false,
-  },
-];
-
 interface ImageGridProps {
   section: ArchiveCategory;
-  handleImageClick: (clickedImg: string) => void;
+  handleImageClick: (clickedImg: string, orientation: "vertical" | "horizontal") => void;
   handleScroll: (e: React.UIEvent<HTMLElement>) => void;
 }
 
@@ -189,7 +50,7 @@ export default function AdvancedGridList(props: ImageGridProps) {
           <GridListTile key={tile.img} cols={tile.featured ? 2 : 1} rows={tile.featured ? 2 : 1} >
             <Image
               src={tile.img}
-              onClick={() => props.handleImageClick(tile.img)}
+              onClick={() => props.handleImageClick(tile.img, tile.orientation)}
               style={{
                 backgroundColor: "transparent",
               }}
