+++ src/Archive/Lightbox.tsx
@@ -1,6 +1,5 @@
 import React from "react";
 import { Backdrop, IconButton } from "@material-ui/core";
-import EasyToSeeTooltip from "../Utils/EasyToSeeTooltip";
 import { Close } from "@material-ui/icons";
 import Image from "material-ui-image";
 import { COLORS } from "../Utils/COLORS";
@@ -8,32 +7,33 @@ import { COLORS } from "../Utils/COLORS";
 interface LightboxProps {
     clicked: boolean;
     img: string;
+    orientation: "vertical" | "horizontal";
     handleClickedClose: () => void;
 }
 
 export default function Lightbox(props: LightboxProps) {
     return (
-        <Backdrop style={{zIndex: 2000, overflow: "auto"}} open={props.clicked}>
-            <EasyToSeeTooltip title="Close">
-                <IconButton 
-                    className="close-backdrop-icon" 
-                    onClick={props.handleClickedClose} 
-                >
-                    <Close htmlColor={COLORS.mainWhite}/>
-                </IconButton>
-            </EasyToSeeTooltip>
-            <Image 
+        <Backdrop style={{ zIndex: 2000, overflow: "auto", backgroundColor: "#000000bf" }} open={props.clicked}>
+            <IconButton
+                className="close-backdrop-icon"
+                onClick={props.handleClickedClose}
+                title="Close"
+            >
+                <Close htmlColor={COLORS.mainWhite} />
+            </IconButton>
+            <Image
                 src={props.img}
                 style={{
                     position: "",
                     paddingTop: "",
                     backgroundColor: "transparent",
-                    width: "60%",
+                    width: props.orientation === "horizontal" ? "60%" : "",
+                    height: props.orientation === "horizontal" ? "" : "90%",
                     margin: "auto",
                 }}
                 imageStyle={{
-                    height: "",
-                    width: "100%",
+                    height: props.orientation === "horizontal" ? "" : "100%",
+                    width: props.orientation === "horizontal" ? "100%" : "",
                     position: "",
                     border: `0.5rem solid ${COLORS.mainWhite}`,
                     borderRadius: "1rem",
