+++ src/Main/Main.scss
@@ -30,14 +30,3 @@
     text-align: center;
     overflow: hidden;
 }
-
-.scroll-to-top.MuiIconButton-root {
-    position: absolute;
-    right: 2rem;
-    background-color: $school-green;
-    box-shadow: 0px 0px 10px black;
-    &:hover {
-        box-shadow: 0px 0px 10px $school-gold;
-        background-color: $school-green;
-    }
-}
